docs(js): improve jsdoc and docs coverage for phoenix-client eval tests
Fill documentation gaps across the CI eval test surface: a file-level
overview and per-export JSDoc for the cross-process report-artifacts
bridge, field docs on the public Annotation/test config/args types and
the SuiteSummary shape, doc comments on the suite-stack and run helpers,
and the missing test.only/test.skip JSDoc. Point the README eval section
at the fuller docs/ pages.
