+++ js/packages/phoenix-client/src/testing/report-artifacts.ts
@@ -1,3 +1,17 @@
+/**
+ * Cross-process bridge for the end-of-run Phoenix summary.
+ *
+ * Both Vitest and Jest run test files in separate worker processes, so a
+ * suite's results live in a different process than the reporter that prints
+ * the final summary. To bridge that gap, each worker writes a small JSON
+ * artifact describing its suite to a shared temp directory when the suite
+ * finishes (see {@link writeSuiteSummaryArtifact}). At the end of the run the
+ * reporter — back in the main process — reads every artifact written since the
+ * run began (see {@link readSuiteSummaryArtifacts}) and prints them together.
+ *
+ * Everything here is best-effort: a failure to write or read an artifact must
+ * never fail a user's test, so the file system calls swallow their errors.
+ */
 import { createHash } from "node:crypto";
 import {
   existsSync,
@@ -15,18 +29,32 @@ import { join } from "node:path";
 import type { SuiteSummary } from "./reporter-format";
 import type { SuiteState, TestResult } from "./state";
 
+/**
+ * Env var that overrides where suite summary artifacts are written. When unset,
+ * a stable per-project subdirectory of the OS temp dir is used (see
+ * {@link getSuiteSummaryReportDir}).
+ */
 export const PHOENIX_TEST_REPORT_DIR_ENV_VAR = "PHOENIX_TEST_REPORT_DIR";
 
 const ARTIFACT_VERSION = 1;
 const ARTIFACT_FILE_PREFIX = "suite-";
 let artifactCounter = 0;
 
+/** The on-disk shape of a single artifact file. */
 interface SuiteSummaryArtifact {
   version: typeof ARTIFACT_VERSION;
   createdAtMs: number;
   suite: SuiteSummary;
 }
 
+/**
+ * Resolve the directory where suite summary artifacts are written and read.
+ *
+ * Honors {@link PHOENIX_TEST_REPORT_DIR_ENV_VAR} when set. Otherwise it derives
+ * a directory under the OS temp dir, namespaced by a hash of the current
+ * working directory so that concurrent test runs in different projects don't
+ * read each other's artifacts.
+ */
 export function getSuiteSummaryReportDir(): string {
   const configuredReportDir = process.env[PHOENIX_TEST_REPORT_DIR_ENV_VAR];
   if (configuredReportDir && configuredReportDir.trim().length > 0) {
@@ -39,6 +67,11 @@ export function getSuiteSummaryReportDir(): string {
   return join(tmpdir(), "phoenix-client-test-reports", cwdHash);
 }
 
+/**
+ * Delete any artifacts left over from a previous run. The reporter calls this
+ * at the start of every run so summaries from an earlier watch-mode invocation
+ * don't bleed into the next one.
+ */
 export function clearSuiteSummaryArtifacts(): void {
   const reportDir = getSuiteSummaryReportDir();
   mkdirSync(reportDir, { recursive: true });
@@ -54,6 +87,14 @@ export function clearSuiteSummaryArtifacts(): void {
   }
 }
 
+/**
+ * Write one suite's summary to the report directory. Called from each worker
+ * when a suite finishes.
+ *
+ * The file is written to a `.tmp` path first and then atomically renamed into
+ * place, so the reader never observes a half-written artifact. Failures are
+ * intentionally swallowed — reporting must not break the user's tests.
+ */
 export function writeSuiteSummaryArtifact(suite: SuiteState): void {
   try {
     const reportDir = getSuiteSummaryReportDir();
@@ -73,6 +114,15 @@ export function writeSuiteSummaryArtifact(suite: SuiteState): void {
   }
 }
 
+/**
+ * Read back all suite summaries the workers wrote during this run, sorted by
+ * suite name for stable output.
+ *
+ * @param sinceMs - When provided, only artifacts created at or after this
+ *   timestamp are returned, so summaries from a previous run that share the
+ *   report directory are ignored. A one-second slack is allowed to absorb
+ *   small clock differences between worker processes.
+ */
 export function readSuiteSummaryArtifacts({
   sinceMs,
 }: {
@@ -115,6 +165,11 @@ export function readSuiteSummaryArtifacts({
   );
 }
 
+/**
+ * Project the live in-memory {@link SuiteState} down to the serializable
+ * {@link SuiteSummary} that gets written to disk. Drops anything that can't or
+ * shouldn't cross the process boundary (clients, tracers, raw Error objects).
+ */
 function createSuiteSummary(suite: SuiteState): SuiteSummary {
   return {
     name: suite.name,
@@ -152,6 +207,11 @@ function createTestResultSummary(result: TestResult): TestResult {
   };
 }
 
+/**
+ * Make a recorded test output safe to JSON-serialize: `bigint` values are
+ * stringified, and anything that still can't be serialized (cycles, etc.)
+ * falls back to its `String()` form rather than throwing.
+ */
 function toJsonSafeValue(value: unknown): unknown {
   if (value === undefined) {
     return undefined;
@@ -167,6 +227,10 @@ function toJsonSafeValue(value: unknown): unknown {
   }
 }
 
+/**
+ * Validate a parsed artifact before trusting it. Guards against version skew
+ * and partially-written or unrelated JSON files in the report directory.
+ */
 function isSuiteSummaryArtifact(value: unknown): value is SuiteSummaryArtifact {
   if (!isRecord(value)) {
     return false;
